Update backends/console.js
Although console.js works fine on its own the aggregation part for counter-type data is not followed similarly in graphite.js, the other backend. That is why, it might be easier to supply the correct values from the source, and keep the aggregation task limited to the stats.js terminal, and not extend it to the backends.