fixed default retention suggestion
the previous settings creates an aggregation problem because of http://graphite.readthedocs.org/en/0.9.10/whisper.html#archives-retention-and-precision

each metric bucket size needs to divide evenly into the next resolution up.  the first bucket (10s:6h) is 2160 elements, so either 1min:6d needs to be the next bucket, and 10min:1800d the following, or the first bucket needs to be changed to 10s:252m.