Update helpers.js to allow any value for sets
Previously any non-numeric value was rejected. This did not allow for adding things like GUIDs or other string values.

I'm not sure of the performance issues here. Perhaps a simple limit on the length of the value would be good, since we don't want tweets injected into the Set. :D