account for double flush timeouts in docker environments
Issue #685 highlighted a problem which still to this day only seems to
appear in docker environments. However, given the numbers extracted
from a container they can be recreated when pushed through the
getFlushTimeout function.

The issue resulted in a timeout value being calculated which was
around 30ms rather than 10s (in the default settings
scenario). @aonischenko contributed a nifty workaround which fixes the
timeout in these scenarios without too much extra code to get your
head around.

The result was run in a docker container for 8 hours, before this
patch the issue was very prevelant and easy to spot and now it is
solved with no clock slippage.